Add files via upload
This is a simple script I created that allows me to organise client data during a pentest or before when I start getting information regarding the engagement. This helps keep me organised. 
I would like to extend this script to do additional things or make it more interactive, say perhaps run at the end of the engagement to compress all the files and encrypt for easy storage and what not.